# web/swagger.yaml
openapi: 3.1.0
info:
  title: L.S.D API
  description: |
    # Large Search of Data API
    
    High-performance API for searching and managing billions of records across PostgreSQL and ClickHouse.
    
    ## Authentication
    
    The API supports two authentication methods:
    
    ### 1. JWT Bearer Token
    Include your JWT token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ### 2. API Key
    Include your API key in the X-API-Key header:
    ```
    X-API-Key: lsd_live_your_api_key_here
    ```
    
    ## Rate Limits
    
    | Client Type | Rate Limit |
    |-------------|------------|
    | Anonymous | 100 req/min |
    | Authenticated User | 1,000 req/min |
    | AI Agent (API Key) | 5,000 req/min |
    
    ## Scopes
    
    - `read` - Read access to tables and records
    - `write` - Write access (future use)
    - `search` - Search across all tables
    - `pipeline` - Access to pipeline endpoints
    - `admin` - Full administrative access
    
  version: 2.0.0
  contact:
    name: L.S.D Support
    url: https://github.com/vinayyadav36/acid
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key with prefix lsd_live_
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Authentication required"
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: Create a new user account with email, username, and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  accessToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid request or password requirements not met
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login to your account
      description: Authenticate with email/username and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  accessToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
          headers:
            Set-Cookie:
              description: HTTP-only cookies for access_token and refresh_token
              schema:
                type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user info
      description: Returns the authenticated user's profile information
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/api-keys:
    get:
      tags: [Authentication]
      summary: List all API keys
      description: Get all API keys for the authenticated user
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKeys:
                    type: array
                    items:
                      $ref: '#/components/schemas/APIKey'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Authentication]
      summary: Create a new API key
      description: Generate a new API key with specified scopes and rate limit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
      responses:
        '200':
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  apiKey:
                    type: object
                    properties:
                      id:
                        type: string
                      key:
                        type: string
                        description: The full API key (only shown once!)
                      keyPrefix:
                        type: string
                      name:
                        type: string
                      scopes:
                        type: array
                        items:
                          type: string
                      rateLimit:
                        type: integer
                  warning:
                    type: string
        '400':
          description: Invalid scopes or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/api-keys/{id}:
    delete:
      tags: [Authentication]
      summary: Revoke an API key
      description: Revoke (delete) an API key by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API key ID
      responses:
        '200':
          description: API key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════
  # HEALTH CHECK
  # ═══════════════════════════════════════════════════════════
...
  /api/tables/{table}/schema:
    get:
      tags: [Tables]
      summary: Get table schema
      description: Get detailed schema information for a specific table
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
          description: Table name
      responses:
        '200':
          description: Table schema
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  schema:
                    type: string
                  columns:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        dataType:
                          type: string
                        nullable:
                          type: boolean
                  primary_key:
                    type: array
                    items:
                      type: string
                  indexes:
                    type: array
                    items:
                      type: object
        '404':
          description: Table not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{table}/records:
    get:
      tags: [Records]
      summary: Get records from a table
      description: Retrieve records from a table with pagination, filtering, and sorting
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor
        - name: sort_by
          in: query
          schema:
            type: string
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  next_cursor:
                    type: string
                  has_more:
                    type: boolean
                  count:
                    type: integer
                  table:
                    type: string
                  cached:
                    type: boolean
  /api/tables/{table}/records/{pk}:
    get:
      tags: [Records]
      summary: Get a single record by primary key
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
        - name: pk
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Single record
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Record not found
  /api/tables/{table}/search:
    get:
      tags: [Search]
      summary: Search within a table
      description: Full-text search within a specific table
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
          description: Search term
        - name: columns
          in: query
          schema:
            type: string
          description: Comma-separated list of columns to search
        - name: engine
          in: query
          schema:
            type: string
            enum: [clickhouse, postgresql]
          description: Force a specific search engine
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  next_cursor:
                    type: string
                  has_more:
                    type: boolean
                  count:
                    type: integer
                  table:
                    type: string
                  search_columns:
                    type: array
                    items:
                      type: string
                  search_engine:
                    type: string
                    enum: [clickhouse, postgresql]
  /api/tables/{table}/stats:
    get:
      tags: [Tables]
      summary: Get table statistics
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Table statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  table_name:
                    type: string
                  estimated_rows:
                    type: integer
                  table_size:
                    type: string
                  index_size:
                    type: string
                  clickhouse_indexed:
                    type: integer
                  clickhouse_last_sync:
                    type: string
                    format: date-time

  # ═══════════════════════════════════════════════════════════
  # GLOBAL SEARCH
  # ═══════════════════════════════════════════════════════════
  /api/search/:
    get:
      tags: [Search]
      summary: Global search across all tables
      description: |
        Optimized full-text search across all indexed tables using ClickHouse bitmap indexing.
        This is the fastest way to search billions of records.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
          description: Search term
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor
      responses:
        '200':
          description: Search results grouped by table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        '503':
          description: ClickHouse not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/search:
    get:
      tags: [Search]
      summary: Global search alias
      description: Same as `/api/search/` for clients that call the non-trailing-slash route.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Search results grouped by table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
  /api/smart-search:
    get:
      tags: [Search]
      summary: Smart classified search
      description: Classifies the query (ID, phone, account, text, etc.) and searches the best matching entity paths.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: scope
          in: query
          schema:
            type: string
            enum: [database, column, row]
            default: database
        - name: dataSourceId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Smart-search result payload
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Missing/invalid query
        '501':
          description: Feature disabled when DB search layer is not enabled
...
  /api/admin/db-search/sources:
    get:
      tags: [Search]
      summary: List DB search sources (admin)
      responses:
        '200':
          description: Search source stats
          content:
            application/json:
              schema:
                type: object
        '403':
          description: Admin role required
...
  /api/pipeline/jobs/{job_id}/stream:
    get:
      tags: [Pipeline]
      summary: Stream job progress
      description: Server-sent events stream for real-time job progress
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string

  # ═══════════════════════════════════════════════════════════
  # CDC ENDPOINTS
  # ═══════════════════════════════════════════════════════════
...